+++ core/dfinscmp.h
@@ -39,7 +39,7 @@
 #include <stdio.h>
 
    void                           SetupDefinstancesCompiler(Environment *);
-   void                           DefinstancesCModuleReference(Environment *,FILE *,int,int,int);
+   void                           DefinstancesCModuleReference(Environment *,FILE *,unsigned long,unsigned int,unsigned int);
 
 #endif /* DEFINSTANCES_CONSTRUCT && CONSTRUCT_COMPILER && (! RUN_TIME) */
